fix(登录): 修复状态比较和重构限流检查逻辑
修复 UserEntity.getStatus() 返回 Integer 对象时使用 == 比较可能失败的问题，改为使用 intValue()。
重构登录限流检查逻辑，提取重复代码为 checkRateLimit 方法，提高代码复用性和可维护性。

+++ backend/src/main/java/elexvx/admin/service/auth/AuthLoginService.java
@@ -113,7 +113,7 @@ public class AuthLoginService {
       log.warn("LOGIN_CREDENTIAL_INVALID reason=PASSWORD_MISMATCH account={} clientIp={}", account, getClientIp());
       throw LoginCredentialException.passwordMismatch(account);
     }
-    if (user.getStatus() != null && user.getStatus() == 0) {
+    if (user.getStatus() != null && user.getStatus().intValue() == 0) {
       log.warn("LOGIN_ACCOUNT_DISABLED account={} clientIp={}", account, getClientIp());
       throw LoginAccountDisabledException.disabled(account);
     }
@@ -291,7 +291,7 @@ public class AuthLoginService {
   }
 
   private void ensureUserActive(UserEntity user) {
-    if (user.getStatus() != null && user.getStatus() == 0) {
+    if (user.getStatus() != null && user.getStatus().intValue() == 0) {
       throw badRequest("账号已禁用");
     }
   }
